client/requirements.txt
Django==3.2
djangorestframework==3.14.0
djangorestframework_jwt==1.11.0
